+++ .github/workflows/ci-workflow.yml
@@ -1,6 +1,6 @@
 name: Build and Test
 
-on: 
+on:
   pull_request:
   push:
     branches:
@@ -11,10 +11,11 @@ env:
 
 jobs:
   test:
-    runs-on: windows-latest
     strategy: 
       matrix: 
         rust: [1.42.0, stable]
+        os: [windows-latest, ubuntu-latest]
+    runs-on: ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v2
     - uses: actions-rs/toolchain@v1
@@ -25,8 +26,8 @@ jobs:
         components: rustfmt
 
     - name: tests
-      run: cargo test --all 
+      run: cargo test --all
 
     - name: fmt
-      run: cargo fmt --all -- --check
       if: matrix.rust == 'stable'
+      run: cargo fmt --all -- --check
