+++ examples/basic/server/src/lib.rs
@@ -1,3 +1,5 @@
+#![cfg(windows)]
+
 mod british_short_hair_cat;
 
 use british_short_hair_cat::BritishShortHairCat;
